SAA-348: Changed ints to longs and integer to bigint for all foreign keys

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsactivitiesmanagementapi/entity/Appointment.kt
@@ -34,7 +34,7 @@ data class Appointment(
 
   var prisonCode: String,
 
-  var internalLocationId: Int?,
+  var internalLocationId: Long?,
 
   var inCell: Boolean,
